Many Alterations.

+++ battle.h
@@ -9,9 +9,9 @@ float battlePortalCoverage( int id, int tx, int ty, int research );
 
 int battleReadinessLoss( dbUserMainPtr maind, dbUserMainPtr main2d );
 
-void battlePhaseUpdate( int *unit, int *results );
+void battlePhaseUpdate( int64_t *unit, int64_t *results );
 
-int battle( int id, int fltid, int *results );
+int battle( int id, int fltid, int64_t *results );
 
 
 #endif
